moving utilities of handling podspec to podspec_helper
