+++ pkg/kn/commands/service/create_mock_test.go
@@ -279,7 +279,7 @@ func TestServiceCreateWithMountConfigMap(t *testing.T) {
 	template := &service.Spec.Template
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: servinglib.GenerateVolumeName("/mount/path"),
+			Name: util.GenerateVolumeName("/mount/path"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -292,7 +292,7 @@ func TestServiceCreateWithMountConfigMap(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      servinglib.GenerateVolumeName("/mount/path"),
+			Name:      util.GenerateVolumeName("/mount/path"),
 			MountPath: "/mount/path",
 			ReadOnly:  true,
 		},
@@ -359,7 +359,7 @@ func TestServiceCreateWithMountSecret(t *testing.T) {
 	template := &service.Spec.Template
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: servinglib.GenerateVolumeName("/mount/path"),
+			Name: util.GenerateVolumeName("/mount/path"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name",
@@ -370,7 +370,7 @@ func TestServiceCreateWithMountSecret(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      servinglib.GenerateVolumeName("/mount/path"),
+			Name:      util.GenerateVolumeName("/mount/path"),
 			MountPath: "/mount/path",
 			ReadOnly:  true,
 		},
