+++ pkg/kn/commands/service/service_update_mock_test.go
@@ -1069,7 +1069,7 @@ func TestServiceUpdateWithAddingMount(t *testing.T) {
 
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: clientserving.GenerateVolumeName("/mount/config-map-path"),
+			Name: util.GenerateVolumeName("/mount/config-map-path"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -1079,7 +1079,7 @@ func TestServiceUpdateWithAddingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/secret-path"),
+			Name: util.GenerateVolumeName("/mount/secret-path"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name",
@@ -1090,12 +1090,12 @@ func TestServiceUpdateWithAddingMount(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/config-map-path"),
+			Name:      util.GenerateVolumeName("/mount/config-map-path"),
 			MountPath: "/mount/config-map-path",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/secret-path"),
+			Name:      util.GenerateVolumeName("/mount/secret-path"),
 			MountPath: "/mount/secret-path",
 			ReadOnly:  true,
 		},
@@ -1134,7 +1134,7 @@ func TestServiceUpdateWithUpdatingMount(t *testing.T) {
 
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: clientserving.GenerateVolumeName("/mount/config-map-path"),
+			Name: util.GenerateVolumeName("/mount/config-map-path"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -1144,7 +1144,7 @@ func TestServiceUpdateWithUpdatingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/secret-path"),
+			Name: util.GenerateVolumeName("/mount/secret-path"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name-1",
@@ -1155,12 +1155,12 @@ func TestServiceUpdateWithUpdatingMount(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/config-map-path"),
+			Name:      util.GenerateVolumeName("/mount/config-map-path"),
 			MountPath: "/mount/config-map-path",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/secret-path"),
+			Name:      util.GenerateVolumeName("/mount/secret-path"),
 			MountPath: "/mount/secret-path",
 			ReadOnly:  true,
 		},
@@ -1174,7 +1174,7 @@ func TestServiceUpdateWithUpdatingMount(t *testing.T) {
 
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: clientserving.GenerateVolumeName("/mount/config-map-path"),
+			Name: util.GenerateVolumeName("/mount/config-map-path"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -1184,7 +1184,7 @@ func TestServiceUpdateWithUpdatingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/secret-path"),
+			Name: util.GenerateVolumeName("/mount/secret-path"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name-2",
@@ -1195,12 +1195,12 @@ func TestServiceUpdateWithUpdatingMount(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/config-map-path"),
+			Name:      util.GenerateVolumeName("/mount/config-map-path"),
 			MountPath: "/mount/config-map-path",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/secret-path"),
+			Name:      util.GenerateVolumeName("/mount/secret-path"),
 			MountPath: "/mount/secret-path",
 			ReadOnly:  true,
 		},
@@ -1243,7 +1243,7 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: clientserving.GenerateVolumeName("/mount/config-map-path-1"),
+			Name: util.GenerateVolumeName("/mount/config-map-path-1"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -1253,7 +1253,7 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/secret-path-1"),
+			Name: util.GenerateVolumeName("/mount/secret-path-1"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name-1",
@@ -1261,7 +1261,7 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/config-map-path-2"),
+			Name: util.GenerateVolumeName("/mount/config-map-path-2"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -1271,7 +1271,7 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/secret-path-2"),
+			Name: util.GenerateVolumeName("/mount/secret-path-2"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name-2",
@@ -1292,22 +1292,22 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/config-map-path-1"),
+			Name:      util.GenerateVolumeName("/mount/config-map-path-1"),
 			MountPath: "/mount/config-map-path-1",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/secret-path-1"),
+			Name:      util.GenerateVolumeName("/mount/secret-path-1"),
 			MountPath: "/mount/secret-path-1",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/config-map-path-2"),
+			Name:      util.GenerateVolumeName("/mount/config-map-path-2"),
 			MountPath: "/mount/config-map-path-2",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/secret-path-2"),
+			Name:      util.GenerateVolumeName("/mount/secret-path-2"),
 			MountPath: "/mount/secret-path-2",
 			ReadOnly:  true,
 		},
@@ -1326,7 +1326,7 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 
 	template.Spec.Volumes = []corev1.Volume{
 		{
-			Name: clientserving.GenerateVolumeName("/mount/config-map-path-1"),
+			Name: util.GenerateVolumeName("/mount/config-map-path-1"),
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
@@ -1336,7 +1336,7 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 			},
 		},
 		{
-			Name: clientserving.GenerateVolumeName("/mount/secret-path-2"),
+			Name: util.GenerateVolumeName("/mount/secret-path-2"),
 			VolumeSource: corev1.VolumeSource{
 				Secret: &corev1.SecretVolumeSource{
 					SecretName: "secret-name-2",
@@ -1357,12 +1357,12 @@ func TestServiceUpdateWithRemovingMount(t *testing.T) {
 
 	template.Spec.Containers[0].VolumeMounts = []corev1.VolumeMount{
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/config-map-path-1"),
+			Name:      util.GenerateVolumeName("/mount/config-map-path-1"),
 			MountPath: "/mount/config-map-path-1",
 			ReadOnly:  true,
 		},
 		{
-			Name:      clientserving.GenerateVolumeName("/mount/secret-path-2"),
+			Name:      util.GenerateVolumeName("/mount/secret-path-2"),
 			MountPath: "/mount/secret-path-2",
 			ReadOnly:  true,
 		},
